chore(sentinel): proc macro to add ToBytes Function
At first i thought it would be a better idea to implement a proc macro
for this. However, identifying types using AST is somewhat tricky +
getting the type to check if it is an enum, and if the enum has a repr
field makes it very hard to go for this approach.

It would be better if we represent as a byte array. That way it is
easier for us.

TODO:
update the enum's to be represented as byte array.
reduce the complexity of the byte array.